fix: Fire ImageOpened event on SetSource[Async]
