+++ src/Uno.UI/UI/Xaml/Media/Imaging/SvgImageSource.wasm.cs
@@ -87,10 +87,6 @@ namespace Windows.UI.Xaml.Media.Imaging
 
 		}
 
-		internal override void ReportImageLoaded() => RaiseImageOpened();
-
-		internal override void ReportImageFailed(string errorMessage) => RaiseImageFailed(SvgImageSourceLoadStatus.Other);
-
 		public override string ToString()
 		{
 			if (WebUri is { } uri)
